feat: 同步缓存点赞数

+++ model/article.go
@@ -8,7 +8,7 @@ type Article struct {
 	Title     string `gorm:"column:title;type:varchar(64);not null;comment:'标题'"`
 	UserID    int64  `gorm:"column:userid;type:bigint;not null;comment:'用户id'"`
 	User      User   `gorm:"foreignKey:UserID;references:ID"`
-	DiggCount int64  `gorm:"column:digg_count;type:bigint;not null;comment:'点赞数'"`
+	DiggCount int    `gorm:"column:digg_count;type:bigint;not null;comment:'点赞数'"`
 }
 
 func (a *Article) TableName() string {
